Move function to store framebuffer to image
